backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
deepface==0.0.93
tf-keras==2.16.0
psycopg2-binary==2.9.9
pgvector==0.3.2
sqlalchemy==2.0.36
python-multipart==0.0.12
python-dotenv==1.0.1
numpy==1.26.4
Pillow==10.4.0
requests==2.32.3
locust==2.32.3
